requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
alembic==1.16.4
python-dotenv>=1.0.0
PyJWT==2.10.1
PyMySQL==1.1.2
mysql-connector-python==9.4.0
psycopg2-binary>=2.9.0
pandas==2.1.4
openpyxl==3.1.2
python-multipart==0.0.6
loguru>=0.7.0
brasilapi==0.1.0
supabase>=2.0.0
requests>=2.31.0
pydantic>=2.0.0
